+++ src/js/policies/firefox.js
@@ -2195,6 +2195,25 @@ const policies = {
         ]
       },
       {
+        name : 'HarmfulAddon',
+        label : browser.i18n.getMessage('policy_description_EnableTrackingProtection_HarmfulAddon'),
+        type : 'enum',
+        options : [
+          {
+            label : browser.i18n.getMessage('enum_value_no_preference'),
+            value : null
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_block_yes'),
+            value : 'true'
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_block_no'),
+            value : 'false'
+          }
+        ]
+      },
+      {
         name : 'BaselineExceptions',
         label : browser.i18n.getMessage('policy_description_EnableTrackingProtection_BaselineExceptions'),
         type : 'enum',